added function pointers to activation and learning functions

+++ run.sh
@@ -1,3 +1,4 @@
-rm core*
+[ -e core*] && rm core*
+[ -e Neural.out] && rm Neural.out
 gcc -o Neural.out *.c -lm -g
-./Neural
+./Neural.out
